# Conflicts:
#	app/Models/Base/RequestType.php
#	app/Models/User.php

+++ routes/api.php
@@ -15,15 +15,13 @@ Route::middleware('throttle:60,1')->group(function () {
     });
 
     Route::middleware('JwtMiddleware')->group(function () {
-        Route::controller(\App\Http\Controllers\Api\RayvarzController::class)->prefix('/rayvarz')->group(function () {
-            // TODO: add middleware
-            Route::post('/sync/{module}/{model_name}', 'sync');
-        });
+        // Route::controller(\App\Http\Controllers\Api\RayvarzController::class)->prefix('/rayvarz')->group(function () {
+        //     Route::post('/sync/{module}/{model_name}', 'sync');
+        // });
 
-        Route::controller(\App\Http\Controllers\Api\KasraController::class)->prefix('/kasra')->group(function () {
-            // TODO: add middleware
-            Route::post('/sync', 'sync');
-        });
+        // Route::controller(\App\Http\Controllers\Api\KasraController::class)->prefix('/kasra')->group(function () {
+        //     Route::post('/sync', 'sync');
+        // });
 
         Route::controller(\App\Http\Controllers\Commerce\TenderController::class)->prefix('/commerce/tender')->group(function () {
             Route::get('/get-actives', 'getActives')->middleware('permission:read Active-Tenders');
@@ -39,29 +37,56 @@ Route::middleware('throttle:60,1')->group(function () {
         });
 
         Route::controller(\App\Http\Controllers\Payroll\PayrollSlipController::class)->prefix('/payroll/payroll-slip')->group(function () {
-            Route::get('/get-the-last-few-months', 'getTheLastFewMonths');
+            Route::get('/get-the-last-few-months', 'getTheLastFewMonths')->middleware('role:Super Admin|employee');
+            // Route::get('print', 'print')->middleware('role:Super Admin|employee');
+            Route::get('reports', 'getReports')->middleware('permission:read Payroll-Batches');
         });
 
         Route::controller(\App\Http\Controllers\PersonnelRecords\PersonnelRecordsController::class)->prefix('/personnel-records')->group(function () {
             Route::get('/get-by-personnel_code', 'getPersonnelRecords')->middleware('permission:read Personnel-Records');
         });
 
-        // TODO: add middleware
         Route::controller(\App\Http\Controllers\Base\UserController::class)->prefix('/base/user')->group(function () {
-            Route::get('/approval-flows-as-requester', 'getApprovalFlowsAsRequester');
+            Route::get('/approval-flows-as-requester', 'getApprovalFlowsAsRequester')->middleware('permission:read Approval-Flows');
         });
 
-        // TODO: add middleware
         Route::controller(\App\Http\Controllers\Base\ApprovalFlowController::class)->prefix('/base/approval-flow')->group(function () {
-            Route::post('/update', 'update'); //->middleware('permission:Edit Approval-Flows');
+            Route::post('/update', 'update')->middleware('permission:edit Approval-Flows');
         });
 
-        // TODO: add middleware
         Route::controller(\App\Http\Controllers\Survey\SurveyController::class)->prefix('/survey/survey')->group(function () {
-            Route::post('/create', 'create');
-            Route::post('/update', 'update');
-            Route::delete('/', 'delete');
-            Route::post('/participate', 'participate');
+            Route::post('/create', 'create')->middleware('permission:read Surveys');
+            Route::post('/update', 'update')->middleware('permission:read Surveys');
+            Route::delete('/', 'delete')->middleware('permission:read Surveys');
+            Route::post('/participate', 'participate')->middleware('role:Super Admin|employee');
+        });
+
+        //Birthday Routes
+        Route::prefix('birthday')->group(function () {
+            Route::controller(\App\Http\Controllers\Birthday\BirthdayGiftController::class)->prefix('gift')->group(function () {
+                Route::post('/', 'store')->middleware('permission:read Birthdays');
+                Route::get('/', 'index')->middleware('permission:read Birthdays');
+                Route::get('/get-actives', 'getActives')->middleware('role:Super Admin|employee');
+                Route::post('/{id}', 'update')->middleware('permission:read Birthdays');
+                Route::delete('/{id}', 'delete')->middleware('permission:read Birthdays');
+            });
+
+            Route::controller(\App\Http\Controllers\Birthday\BirthdayFileController::class)->prefix('file')->group(function () {
+                Route::post('/', 'store')->middleware('permission:read Birthdays');
+                Route::get('/', 'index')->middleware('permission:read Birthdays');
+                Route::post('/{id}', 'update')->middleware('permission:read Birthdays');
+                Route::delete('/{id}', 'delete')->middleware('permission:read Birthdays');
+                Route::get('/statistics', 'statistics')->middleware('permission:read Birthdays');
+            });
+
+            Route::controller(\App\Http\Controllers\Birthday\BirthdayFileUserController::class)->prefix('user')->group(function () {
+                Route::post('/', 'store')->middleware('permission:read Birthdays');
+                Route::get('/', 'index')->middleware('permission:read Birthdays');
+                Route::delete('/delete', 'delete')->middleware('permission:read Birthdays');
+                Route::post('/status', 'changeStatus')->middleware('permission:read Birthdays');
+                Route::post('/choose', 'chooseBirthdayGift')->middleware('role:Super Admin|employee');
+                Route::get('/check', 'checkAccess')->middleware('role:Super Admin|employee');
+            });
         });
 
         // TODO: add middleware
